requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-WTF==1.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
scikit-learn==1.3.0
scipy==1.11.2
threadpoolctl==3.2.0
Werkzeug==2.2.3
WTForms==3.0.1
